requirements.txt
streamlit==1.40.1 
langchain==0.3.7
langchain-openai
langchain-azure-dynamic-sessions
scipy==1.14.1
scikit-learn==1.14.1
numpy==1.26.4
Pillow
boto3==1.35.35
neo4j==5.26.0
pandas==2.2.3
streamlit_tree_select
openai==1.14.2
tiktoken
cohere
streamlit_cookies_manager
pathlib